Little refactor (#103)
* Fixes

* Variable name

* Unused method fix

* Requested changes

* Requested changes

* Fix

* Fix

* Best practices exceptions
