+++ .reek.yml
@@ -18,7 +18,7 @@ detectors:
     min_clump_size: 2
   DuplicateMethodCall:
     enabled: true
-    exclude: []
+    exclude: ['GithubService#handle_event']
     max_calls: 1
     allow_calls: []
   FeatureEnvy:
