--- app/models/concerns/event_helper.rb
@@ -1,7 +1,0 @@
-##
-# This Module adds common method to all ApplicationRecord modeling Events
-module EventHelper
-  def event_name
-    self.class.name.demodulize.underscore
-  end
-end
