+++ app/models/events/pull_request.rb
@@ -30,8 +30,6 @@
 #
 module Events
   class PullRequest < ApplicationRecord
-    include EventHelper
-
     enum state: { open: 'open', closed: 'closed' }
 
     belongs_to :project, inverse_of: :pull_requests
