+++ app/services/event_builders/pull_request.rb
@@ -7,14 +7,8 @@ module EventBuilders
     def build
       pr_data = @payload['pull_request']
       Events::PullRequest.find_or_create_by!(github_id: pr_data['id']) do |pr|
-        pr.project = find_or_create_project(@payload['repository'])
+        pr.project = Projects::Builder.fetch_or_create(@payload['repository'])
         ATTR_PAYLOAD_MAP.each { |key, value| pr.public_send("#{key}=", pr_data.fetch(value)) }
-      end
-    end
-
-    def find_or_create_project(repository_data)
-      Project.find_or_create_by!(github_id: repository_data['id']) do |project|
-        project.lang = repository_data['language']
       end
     end
   end
