+++ app/services/github_service.rb
@@ -13,39 +13,34 @@ class GithubService < BaseService
   private
 
   def handle_request
-    build_project
     handle_event
-    raise Events::NotHandleableError unless @entity
-
     handle_action
   end
 
-  def build_project
-    repo = @payload['repository']
-    @project = Project.find_or_create_by!(github_id: repo['id']) do |pj|
-      pj.name = repo['name']
-      pj.description = repo['description']
-    end
-  end
-
   def handle_event
-    @entity = find_or_create_event_type
-
-    Event.create!(project: @project,
-                  data: @payload,
-                  name: @event,
-                  handleable: @entity,
-                  occurred_at: occurence_time(event_name: @event, payload: @payload))
+    project = Projects::Builder.fetch_or_create(@payload['repository'])
+    event = Event.new(project: project, data: @payload, name: @event)
+    if handleable_event?
+      @entity = find_or_create_event_type
+      event.assign_attributes(handleable: @entity, occurred_at: occurence_time)
+      event.save!
+    else
+      event.save!
+      raise Events::NotHandleableError
+    end
+    event
   end
 
   def handleable_event?
-    handleable_event_type?(@event)
+    Event::TYPES.include?(@event)
   end
 
   def find_or_create_event_type
-    return unless handleable_event?
-
     EventBuilders.const_get(@event.classify).call(payload: @payload, event: @event)
+  end
+
+  def occurence_time
+    send("#{@event}_ocurrence_time", payload: @payload)
   end
 
   def handle_action
