+++ lib/time_interval.rb
