+++ lib/time_intervals/daily_interval.rb
