+++ spec/models/event_spec.rb
@@ -29,24 +29,8 @@ RSpec.describe Event, type: :model do
     it { should validate_presence_of(:name) }
     it { should validate_presence_of(:data) }
 
-    context 'with an event of a handled type' do
-      subject { build(:event_pull_request, name: 'review') }
-
-      it 'validates the name to match the event type' do
-        expect(subject).not_to be_valid
-      end
-    end
-
     context 'with a pull_request event' do
       subject { create :event_pull_request }
-      let(:pull_request_event_with_invalid_name) { build(:event_pull_request, name: 'review') }
-
-      it { should validate_presence_of(:handleable) }
-      it { should validate_presence_of(:occurred_at) }
-
-      it 'validates the event name to match the event type' do
-        expect(pull_request_event_with_invalid_name).not_to be_valid
-      end
 
       it 'sets the occcured_at attribute from the payload data' do
         expect(subject.occurred_at).not_to be_nil
@@ -55,14 +39,6 @@ RSpec.describe Event, type: :model do
 
     context 'with a review event' do
       subject { create :event_review }
-      let(:review_event_with_invalid_name) { build(:event_review, name: 'pull_request') }
-
-      it { should validate_presence_of(:handleable) }
-      it { should validate_presence_of(:occurred_at) }
-
-      it 'validates the event name to match the event type' do
-        expect(review_event_with_invalid_name).not_to be_valid
-      end
 
       it 'sets the occcured_at attribute from the payload data' do
         expect(subject.occurred_at).not_to be_nil
@@ -71,16 +47,6 @@ RSpec.describe Event, type: :model do
 
     context 'with a review_comment event' do
       subject { create :event_review_comment }
-      let(:review_comment_event_with_invalid_name) do
-        build(:event_review_comment, name: 'pull_request')
-      end
-
-      it { should validate_presence_of(:handleable) }
-      it { should validate_presence_of(:occurred_at) }
-
-      it 'validates the event name to match the event type' do
-        expect(review_comment_event_with_invalid_name).not_to be_valid
-      end
 
       it 'sets the occcured_at attribute from the payload data' do
         expect(subject.occurred_at).not_to be_nil
