+++ spec/services/metrics/review_turnaround_per_project_processor_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../support/metrics_specs_helper'
+require_relative '../../support/helpers/metrics_specs_helper'
 
 RSpec.describe Metrics::ReviewTurnaroundPerProjectProcessor do
   include_context 'events metrics'
