+++ spec/services/metrics/time_intervals/daily_metrics_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../../support/metrics_specs_helper'
+require_relative '../../../support/helpers/metrics_specs_helper'
 
 RSpec.describe Metrics::MetricDefinitionTimeIntervalsProcessor do
   subject { described_class }
